// FUTURE_FIX/@alzhan/rncore-spacing-props-module/hooks/useSpacings/types.ts
import { DimensionValue } from 'react-native'

export interface Spacings {
  marginT?: DimensionValue | undefined
  marginB?: DimensionValue | undefined
  marginL?: DimensionValue | undefined
  marginR?: DimensionValue | undefined

  marginHor?: DimensionValue | undefined
  marginVer?: DimensionValue | undefined

  paddingT?: DimensionValue | undefined
  paddingB?: DimensionValue | undefined
  paddingL?: DimensionValue | undefined
  paddingR?: DimensionValue | undefined

  paddingHor?: DimensionValue | undefined
  paddingVer?: DimensionValue | undefined
}